+++ packages/react-app-super-template/src/types/index.d.ts
@@ -1,3 +1,8 @@
 declare module 'api' {
   export const clientPromise: Promise<any>;
 }
+
+declare module 'api/schema.json' {
+  const _default: any;
+  export default _default;
+}