+++ packages/react-app-super-template/CHANGELOG.md
@@ -1,3 +1,7 @@
+# x.y.z
+
+- Overwrite API schema `server.url` with `config.backend.baseURL`
+
 # 0.1.3
 
 - Fix build